BONUS 1: Write a program to toggle all characters in a string: in a given string, toggle all characters to their opposite case. Special characters will not change.
Got a string from a prompt, I split the string, then a used map() to create a new array and used the elements inside to check if the elements are upperCase() if so, change to lowerCase else leave upperCase.

BONUS 2:
Write a program that takes a given string appended with a number at the end.
The program will check the length of the string to verify the string is equal to the number appended.
If the number appended equals the string length, output "Yes". if the number appended does not equal the string length, output "No".

-got a string from prompt. Got the length of the string. got the number at the end using .match(regex) then checked if length === number